requirements.txt
autopep8==2.0.4
blinker==1.7.0
cffi==1.16.0
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
packaging==23.2
psycopg2==2.9.9
pycodestyle==2.11.1
pycparser==2.21
python-dotenv==1.0.0
tomli==2.0.1
Werkzeug==3.0.1
zipp==3.17.0
